scripts/requirements.txt
atplotlib
https://github.com/matplotlib/mpl_finance/archive/master.zip